pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Backend API for a central Escape Room control system"
authors = ["wwalewski <w.walewski112@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sse-starlette = "^1.6.1"
asyncio = "^3.4.3"
websockets = "^11.0.3"
faker = "^18.10.1"
python-slugify = "^8.0.1"
pyhumps = "^3.8.0"
jsonschema = "^4.18.4"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
pydantic-settings = "^2.0.2"
aiofiles = "^23.1.0"
gmqtt = "^0.6.12"
redis = {extras = ["hiredis"], version = "^4.6.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
pylint = "^2.17.4"
bump-pydantic = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
